Writing and reading by email lowercase Lekcja 21.05.25 - definiowanie liczb pierwszych int sqrt nie działa :(

+++ src/faktoryzacja.py
@@ -4,7 +4,7 @@ n =
 def liczba_pierwsza (n):
     if n<2:
         return False
-    for i in range (2, sqrt(n)):
+    for i in range (2, int(sqrt(n))):
         if n%i == 0:
             return False
     return True
